DNM: adding steps to allow building something
This task requires two things:
- a source QCOW2 image (RHEL iso image converted to QCOW2)
- a certificate and key to sign the QCOW2 image after modification

We did not figure out how to retrieve those at the moment, so I am
adding some code to fake the information in the task, so that we can
at least make the rest of it work.

When we reach a point where we can build *something*, we can figure out
how to fix those.

Signed-off-by: Julien Ropé <jrope@redhat.com>