DNM: Add task for dm-verity build
This is a very specialized workflow, targetting the build of a
Confidential Containers image based on RHEL.

This commit create the task based on buildah-remote-oci-ta task, and naively
applying the steps from https://github.com/confidential-devhub/coco-podvm-scripts/blob/main/example_run.sh

Also turn the needed variables into parameters to the task, so that we can
customize the build from the pipeline definition.

At this stage, I don't expect it to run, but we need to start somewhere to iterate.

Signed-off-by: Julien Ropé <jrope@redhat.com>